+++ .eslintrc.cjs
@@ -5,7 +5,12 @@ require('@rushstack/eslint-patch/modern-module-resolution');
 
 module.exports = {
   root: true,
-  extends: ['airbnb-base/legacy', 'prettier', 'plugin:import/recommended'],
+  extends: [
+    'airbnb-base/legacy',
+    'prettier',
+    'plugin:import/recommended',
+    'plugin:n/recommended',
+  ],
   parser: '@babel/eslint-parser',
   parserOptions: {
     sourceType: 'module',